refactor(paths): calculate paths from mixin
This commit moves the path calculation into the mixin, which
in turn makes the render method mostly redundant. It has
been removed. Props can now be correctly calculated before
being passed to the decorator
